requirements.txt
asyncpg==0.21.0
rich==9.6.0
pyyaml==5.3.1
geoip2==4.0.2
ujson==4.0.2